requirements.txt
aixplain
beautifulsoup4
chromadb
Flask
python-dotenv
requests
